* Fixed major bug where ObjectLists were not being sorted before leaving physics. Added some checks to make sure (mostly) that it doesn't happen again. * Code touchup.

+++ PauseInitiatorID.cpp
@@ -4,7 +4,7 @@
 namespace hg {
 PauseInitiatorID::PauseInitiatorID(pauseinitiatortype::PauseInitiatorType type,
                                    unsigned int ID,
-                                   unsigned int timelineLength) :
+                                   size_t timelineLength) :
 type_(type),
 ID_(ID),
 timelineLength_(timelineLength)
